Implemented loading settings (again), moved ApplicationTest to proper package
