+++ src/test/java/fi/helsinki/cs/tmc/cli/ApplicationTest.java
@@ -1,4 +1,4 @@
-package fi.helsinki.cs.tmc.clii;
+package fi.helsinki.cs.tmc.cli;
 
 import static org.junit.Assert.assertTrue;
